pyproject.toml
[tool.poetry]
name = "setlist-fm-client"
version = "0.1.0"
description = "a python client for the setlist.fm api"
authors = ["Zach Schumacher <zschu15@gmail.com>"]
license = "MIT"
readme = "README.md"
classifiers = [
    'Development Status :: 5 - Production/Stable',
    'Programming Language :: Python',
    'Programming Language :: Python :: 3',
    'Programming Language :: Python :: 3 :: Only',
    'Programming Language :: Python :: 3.7',
    'Programming Language :: Python :: 3.8',
    'Programming Language :: Python :: 3.9',
    'Programming Language :: Python :: 3.10',
    'License :: OSI Approved :: MIT License',
    'Operating System :: Unix',
    'Operating System :: POSIX :: Linux',
    'Environment :: Console',
    'Environment :: MacOS X',
    'Topic :: Software Development :: Libraries :: Python Modules',
]

[tool.poetry.dependencies]
python = ">=3.7,<=3.10"
httpx = "^0.23.0"
pydantic = "^1.9.1"
pyhumps = "^3.7.1"

[tool.poetry.dev-dependencies]
Pygments = "^2.12.0"
pytest-vcr = "^1.0.2"
black = "^22.3.0"
isort = "^5.10.1"
mypy = "^0.961"
pytest-asyncio = "^0.18.3"
pytest-dotenv = "^0.5.2"
devtools = "^0.8.0"
lazydocs = "^0.4.8"
mkdocs = "^1.3.0"
mkdocs-material = "^8.3.8"
pytest-cov = "^3.0.0"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.circleci
    | \.git
    | \.github
    | \.hg
    | \.mypy_cache
    | \.pytest_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | venv
    | media
    | staticfiles
  )/
)
'''


[tool.pytest.ini_options]
norecursedirs = [".git", "venv", ".venv"]
env_files = [".env"]

[tool.coverage.run]
omit = [
    "pydapper/enums.py",
    "tests/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "Protocol",
    "except ImportError",
    "@abstractmethod"
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
